requirements.txt
GitPython==3.1.43
python-dotenv==0.21.1
requests==2.31.0